# openapi.yaml
openapi: 3.0.3
info:
  title: Vokter API
  version: 1.0.0
  description: API de comercio electrónico para clientes web y móvil.
servers:
  - url: http://localhost:5000/api/v1
...
paths:
...
  /auth/register:
    post:
      tags: [Identity]
      summary: Register a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201': {$ref: '#/components/responses/AuthSuccess'}
        '400': {$ref: '#/components/responses/ValidationError'}
        '409': {$ref: '#/components/responses/ConflictError'}
  /auth/login:
    post:
      tags: [Identity]
      summary: Authenticate a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200': {$ref: '#/components/responses/AuthSuccess'}
        '401': {$ref: '#/components/responses/UnauthorizedError'}
  /auth/me:
    get:
      tags: [Identity]
      summary: Get the authenticated user
      security: [{bearerAuth: []}]
      parameters:
        - {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
        - {name: per_page, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
        - {name: status, in: query, schema: {type: string, enum: [pending, paid, processing, shipped, completed, cancelled]}}
      responses:
        '200':
          description: Authenticated user
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: {$ref: '#/components/schemas/User'}
        '401': {$ref: '#/components/responses/UnauthorizedError'}

    patch:
      tags: [Identity]
      summary: Update the authenticated profile
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileUpdateRequest'
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: {$ref: '#/components/schemas/User'}
        '400': {$ref: '#/components/responses/ValidationError'}
  /auth/me/password:
    patch:
      tags: [Identity]
      summary: Change the authenticated user's password
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordChangeRequest'
      responses:
        '204': {description: Password changed}
        '400': {$ref: '#/components/responses/ValidationError'}
        '401': {$ref: '#/components/responses/UnauthorizedError'}
  /auth/refresh:
    post:
      tags: [Identity]
      summary: Rotate a refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Rotated tokens
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: {$ref: '#/components/schemas/TokenPair'}
        '401': {$ref: '#/components/responses/UnauthorizedError'}
  /auth/logout:
    post:
      tags: [Identity]
      summary: Revoke a refresh session
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '204': {description: Refresh session revoked}
        '401': {$ref: '#/components/responses/UnauthorizedError'}
...
  /orders:
    get:
      tags: [Orders]
      summary: List the authenticated user's orders
      security: [{bearerAuth: []}]
      responses:
        '200':
          description: Order list
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: {$ref: '#/components/schemas/Order'}
  /orders/checkout:
    post:
      tags: [Orders]
      summary: Create an order from an active cart
      security: [{bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutRequest'
      responses:
        '201':
          description: Created or previously created order
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: {$ref: '#/components/schemas/Order'}
        '400': {$ref: '#/components/responses/ValidationError'}
        '401': {$ref: '#/components/responses/UnauthorizedError'}
        '409': {$ref: '#/components/responses/ConflictError'}
  /orders/{order_id}:
    get:
      tags: [Orders]
      summary: Get an owned order
      security: [{bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/OrderId'}
      responses:
        '200':
          description: Order
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: {$ref: '#/components/schemas/Order'}
        '404': {$ref: '#/components/responses/NotFoundError'}
  /orders/{order_id}/cancel:
    post:
      tags: [Orders]
      summary: Cancel an owned pending or paid order
      security: [{bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/OrderId'}
      responses:
        '200':
          description: Cancelled order
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: {$ref: '#/components/schemas/Order'}
        '409': {$ref: '#/components/responses/ConflictError'}
  /orders/{order_id}/status:
    patch:
      tags: [Orders]
      summary: Update order status
      security: [{bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/OrderId'}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderStatusRequest'
      responses:
        '200': {$ref: '#/components/responses/OrderSuccess'}
        '403': {$ref: '#/components/responses/ForbiddenError'}
        '400': {$ref: '#/components/responses/ValidationError'}
  /orders/{order_id}/payment:
    patch:
      tags: [Orders]
      summary: Update payment status
      security: [{bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/OrderId'}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentStatusRequest'
      responses:
        '200': {$ref: '#/components/responses/OrderSuccess'}
        '403': {$ref: '#/components/responses/ForbiddenError'}
  /orders/{order_id}/shipment:
    patch:
      tags: [Orders]
      summary: Update shipment status
      security: [{bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/OrderId'}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShipmentStatusRequest'
      responses:
        '200': {$ref: '#/components/responses/OrderSuccess'}
        '403': {$ref: '#/components/responses/ForbiddenError'}
...
  /social/reviews:
    post:
      tags: [Social]
      summary: Create a verified purchase review
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewRequest'
      responses:
        '201': {$ref: '#/components/responses/ReviewSuccess'}
        '400': {$ref: '#/components/responses/ValidationError'}
        '409': {$ref: '#/components/responses/ConflictError'}
  /social/products/{product_id}/reviews:
    get:
      tags: [Social]
      summary: List product reviews
      parameters:
        - {$ref: '#/components/parameters/ProductId'}
        - {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
        - {name: per_page, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 10}}
      responses:
        '200':
          description: Product reviews
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: {$ref: '#/components/schemas/Review'}
        '404': {$ref: '#/components/responses/NotFoundError'}
  /social/reviews/{review_id}:
    patch:
      tags: [Social]
      summary: Update an owned review
      security: [{bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/ReviewId'}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewUpdateRequest'
      responses:
        '200': {$ref: '#/components/responses/ReviewSuccess'}
        '404': {$ref: '#/components/responses/NotFoundError'}
    delete:
      tags: [Social]
      summary: Delete an owned review
      security: [{bearerAuth: []}]
      parameters:
        - {$ref: '#/components/parameters/ReviewId'}
      responses:
        '204': {description: Review deleted}
        '404': {$ref: '#/components/responses/NotFoundError'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ProductId:
      name: product_id
      in: path
      required: true
      schema: {type: integer, minimum: 1}
...
    OrderId:
      name: order_id
      in: path
      required: true
      schema: {type: integer, minimum: 1}
...
    ReviewId:
      name: review_id
      in: path
      required: true
      schema: {type: integer, minimum: 1}
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 1, maxLength: 128}
  schemas:
    RegisterRequest:
      type: object
      required: [email, password, first_name, last_name]
      properties:
        email: {type: string, format: email, maxLength: 255}
        password: {type: string, format: password, minLength: 8}
        first_name: {type: string, maxLength: 100}
        last_name: {type: string, maxLength: 100}
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: {type: string, format: email, maxLength: 255}
        password: {type: string, format: password}
    ProfileUpdateRequest:
      type: object
      properties:
        email: {type: string, format: email, maxLength: 255}
        first_name: {type: string, maxLength: 100}
        last_name: {type: string, maxLength: 100}
    PasswordChangeRequest:
      type: object
      required: [current_password, new_password]
      properties:
        current_password: {type: string, format: password}
        new_password: {type: string, format: password, minLength: 8}
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token: {type: string}
    TokenPair:
      type: object
      required: [access_token, refresh_token]
      properties:
        access_token: {type: string}
        refresh_token: {type: string}
    User:
      type: object
      required: [id, email, first_name, last_name, role, is_active]
      properties:
        id: {type: integer}
        email: {type: string, format: email}
        first_name: {type: string}
        last_name: {type: string}
        role: {type: string, enum: [customer, admin]}
        is_active: {type: boolean}
...
    CheckoutRequest:
      type: object
      required: [cart_id, shipping_address]
      properties:
        cart_id: {type: integer, minimum: 1}
        address_id: {type: integer, minimum: 1, nullable: true}
        shipping_address: {type: string, minLength: 1, maxLength: 500}
        payment_method: {type: string, enum: [card, paypal, cash_on_delivery], default: card}
    Order:
      type: object
      required: [id, user_id, status, total, shipping_address, items, status_history, payment, shipment]
      properties:
        id: {type: integer}
        user_id: {type: integer}
        status: {type: string, enum: [pending, paid, processing, shipped, completed, cancelled]}
        total: {type: string, pattern: '^\d+\.\d{2}$'}
        shipping_address: {type: string}
        items:
          type: array
          items: {$ref: '#/components/schemas/OrderItem'}
        status_history:
          type: array
          items: {$ref: '#/components/schemas/OrderStatusHistory'}
        payment: {$ref: '#/components/schemas/Payment'}
        shipment: {$ref: '#/components/schemas/Shipment'}
    OrderItem:
      type: object
      required: [id, variant_id, product_name, quantity, unit_price]
      properties:
        id: {type: integer}
        variant_id: {type: integer}
        product_name: {type: string}
        quantity: {type: integer, minimum: 1}
        unit_price: {type: string, pattern: '^\d+\.\d{2}$'}
    OrderStatusHistory:
      type: object
      required: [status, created_at]
      properties:
        status: {type: string, enum: [pending, paid, processing, shipped, completed, cancelled]}
        note: {type: string, nullable: true}
        created_at: {type: string, format: date-time}
    Payment:
      type: object
      required: [id, status, method, amount, provider, provider_reference]
      properties:
        id: {type: integer}
        status: {type: string, enum: [pending, authorized, paid, failed, refunded]}
        method: {type: string, enum: [card, paypal, cash_on_delivery]}
        amount: {type: string, pattern: '^\d+\.\d{2}$'}
        provider: {type: string, nullable: true}
        provider_reference: {type: string, nullable: true}
    Shipment:
      type: object
      required: [id, status, tracking_number, carrier]
      properties:
        id: {type: integer}
        status: {type: string, enum: [pending, preparing, shipped, delivered, returned]}
        tracking_number: {type: string, nullable: true}
        carrier: {type: string, nullable: true}
    OrderStatusRequest:
      type: object
      required: [status]
      properties:
        status: {type: string, enum: [pending, paid, processing, shipped, completed, cancelled]}
    PaymentStatusRequest:
      type: object
      required: [status]
      properties:
        status: {type: string, enum: [pending, authorized, paid, failed, refunded]}
    ShipmentStatusRequest:
      type: object
      required: [status]
      properties:
        status: {type: string, enum: [pending, preparing, shipped, delivered, returned]}
        tracking_number: {type: string, maxLength: 120, nullable: true}
        carrier: {type: string, maxLength: 100, nullable: true}
...
    Review:
      type: object
      required: [id, user_id, product_id, rating, is_verified_purchase]
      properties:
        id: {type: integer}
        user_id: {type: integer}
        product_id: {type: integer}
        rating: {type: integer, minimum: 1, maximum: 5}
        title: {type: string, nullable: true}
        content: {type: string, nullable: true}
        is_verified_purchase: {type: boolean}
    CreateReviewRequest:
      type: object
      required: [product_id, rating]
      properties:
        product_id: {type: integer, minimum: 1}
        rating: {type: integer, minimum: 1, maximum: 5}
        title: {type: string, nullable: true}
        content: {type: string, nullable: true}
    ReviewUpdateRequest:
      type: object
      properties:
        rating: {type: integer, minimum: 1, maximum: 5}
        title: {type: string, nullable: true}
        content: {type: string, nullable: true}
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string}
            message: {type: string}
            request_id: {type: string, format: uuid, nullable: true}
  responses:
    AuthSuccess:
      description: Authenticated user and token pair
      content:
        application/json:
          schema:
            type: object
            required: [data]
            properties:
              data:
                allOf:
                  - {$ref: '#/components/schemas/TokenPair'}
                  - type: object
                    required: [user]
                    properties:
                      user: {$ref: '#/components/schemas/User'}
    OrderSuccess:
      description: Order response
      content:
        application/json:
          schema:
            type: object
            required: [data]
            properties:
              data: {$ref: '#/components/schemas/Order'}
...
    ReviewSuccess:
      description: Review response
      content:
        application/json:
          schema:
            type: object
            required: [data]
            properties:
              data: {$ref: '#/components/schemas/Review'}
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    UnauthorizedError:
      description: Authentication required or token invalid
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    ForbiddenError:
      description: Insufficient permissions
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    ConflictError:
      description: Business or data conflict
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}